pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uartl"
version = "0.1.0"
description = "Uart (Serial) based link layer library"
authors = [{ name = "Didas72" }]
dependencies = ["pyserial>=3"]
requires-python = ">= 3.9"
readme = "README.md"
license = "MIT"

[tool.setuptools]
package-dir = {"" = "python"}

[tool.setuptools.packages.find]
where = ["python"]
